Passenger Arrival Distribution: XML stuff
Code to save a passenger distribution to XML and load it in.  Tough luck
if you wanted JSON ;)
Not implemented the code to actually calculate the distribution yet...
Have also changed the main method to test the XML stuff instead of the
lift stuff.  Feel free to comment my new stuff and uncomment the old.