Added a stopping state
It's necessary to differentiate between 'moving' and 'stopping'
(previously both just 'moving') when a car is allocated a hall call for
the floor that the lift is currently moving/stopping at.  If the car is
stopping then we can put it in p1 to be served immediately, if not then
it needs to go in p3 for when the car comes around next time.